Clases TransferDto y TransferMapper creadas. Se eliminaron y/o mvieron paquetes y clases que estaban fuera de lugar, se renombró algún paquete también. Se agregaron métodos de validación de cuentas a TrsanferService

+++ C23650G1/src/main/java/com/cac/C23650G1/entities/Transfer.java
@@ -6,6 +6,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Entity
@@ -33,7 +34,7 @@ public class Transfer {
     private LocalDate date;
 
     @Column(name = "amount")
-    private float amount;
+    private BigDecimal amount;
 
 
 }
